Move rotating caliper functions out of Distance module
These functions are over 700 LOC, and make navigation a chore. This
moves them to a crate-public (but non-public-API) sibling module.
