+++ src/algorithm/polygon_distance_fast_path.rs
@@ -1,0 +1,704 @@
+use num_traits::{Float, Signed};
+use num_traits::float::FloatConst;
+use types::{Point, Polygon};
+use algorithm::distance::Distance;
+use algorithm::extremes::ExtremeIndices;
+
+// These are helper functions for the "fast path" of Polygon-Polygon distance
+// They use the rotating calipers method to speed up calculations.
+// Tests for these functions are in the Distance module
+
+// calculate the minimum distance between two disjoint and linearly separable convex polygons
+pub(crate) fn min_poly_dist<T>(poly1: &Polygon<T>, poly2: &Polygon<T>) -> T
+where
+    T: Float + FloatConst + Signed,
+{
+    let poly1_extremes = poly1.extreme_indices().unwrap();
+    let poly2_extremes = poly2.extreme_indices().unwrap();
+    let ymin1 = poly1.exterior.0[poly1_extremes.ymin];
+    let ymax2 = poly2.exterior.0[poly2_extremes.ymax];
+
+    let mut state = Polydist {
+        poly1: poly1,
+        poly2: poly2,
+        dist: T::infinity(),
+        ymin1: ymin1,
+        ymax2: ymax2,
+        // initial polygon 1 min y idx
+        p1_idx: poly1_extremes.ymin,
+        // initial polygon 2 max y idx
+        q2_idx: poly2_extremes.ymax,
+        // set p1 and q2 to p1ymin and p2ymax initially
+        p1: ymin1,
+        q2: ymax2,
+        p1next: Point::new(T::zero(), T::zero()),
+        q2next: Point::new(T::zero(), T::zero()),
+        p1prev: Point::new(T::zero(), T::zero()),
+        q2prev: Point::new(T::zero(), T::zero()),
+        alignment: None,
+        ap1: T::zero(),
+        aq2: T::zero(),
+        start: None,
+        finished: false,
+        ip1: false,
+        iq2: false,
+        slope: T::zero(),
+        vertical: false,
+    };
+    while !state.finished {
+        nextpoints(&mut state);
+        computemin(&mut state);
+    }
+    state.dist
+}
+
+#[derive(Debug)]
+enum Aligned {
+    EdgeVertexP,
+    EdgeVertexQ,
+    EdgeEdge,
+}
+
+// distance-finding state
+#[derive(Debug)]
+pub(crate) struct Polydist<'a, T>
+where
+    T: Float,
+    T: 'a,
+{
+    poly1: &'a Polygon<T>,
+    poly2: &'a Polygon<T>,
+    dist: T,
+    ymin1: Point<T>,
+    p1_idx: usize,
+    ymax2: Point<T>,
+    q2_idx: usize,
+    p1: Point<T>,
+    q2: Point<T>,
+    p1next: Point<T>,
+    q2next: Point<T>,
+    p1prev: Point<T>,
+    q2prev: Point<T>,
+    alignment: Option<Aligned>,
+    ap1: T,
+    aq2: T,
+    start: Option<bool>,
+    finished: bool,
+    ip1: bool,
+    iq2: bool,
+    slope: T,
+    vertical: bool,
+}
+
+// used to check the sign of a vec of floats
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+enum ListSign {
+    Empty,
+    Positive,
+    Negative,
+    Mixed,
+}
+
+impl<T> Polygon<T>
+where
+    T: Float + Signed,
+{
+    // Is this polygon convex?
+    pub(crate) fn convex(&self) -> bool {
+        let convex = self
+            .exterior
+            .0
+            .iter()
+            .enumerate()
+            .map(|(idx, _)| {
+                let prev_1 = self.prev_vertex(&idx);
+                let prev_2 = self.prev_vertex(&prev_1);
+                cross_prod(&self.exterior.0[prev_2],
+                           &self.exterior.0[prev_1],
+                           &self.exterior.0[idx])
+            })
+            // accumulate and check cross-product result signs in a single pass
+            // positive implies ccw convexity, negative implies cw convexity
+            // anything else implies non-convexity
+            .fold(ListSign::Empty, |acc, n| {
+                match (acc, n.is_positive()) {
+                    (ListSign::Empty, true) | (ListSign::Positive, true) => ListSign::Positive,
+                    (ListSign::Empty, false) | (ListSign::Negative, false) => ListSign::Negative,
+                    _ => ListSign::Mixed
+                }
+            });
+        match convex {
+            ListSign::Mixed => false,
+            _ => true,
+        }
+    }
+    // Wrap-around next and previous Polygon indices
+    fn next_vertex(&self, current_vertex: &usize) -> usize
+    where
+        T: Float,
+    {
+        (current_vertex + 1) % (self.exterior.0.len() - 1)
+    }
+    fn prev_vertex(&self, current_vertex: &usize) -> usize
+    where
+        T: Float,
+    {
+        (current_vertex + (self.exterior.0.len() - 1) - 1) % (self.exterior.0.len() - 1)
+    }
+}
+
+// much of the following code is ported from Java, copyright 1999 Hormoz Pirzadeh, available at:
+// http://web.archive.org/web/20150330010154/http://cgm.cs.mcgill.ca/%7Eorm/rotcal.html
+fn unitvector<T>(slope: &T, poly: &Polygon<T>, p: &Point<T>, idx: &usize) -> Point<T>
+where
+    T: Float + Signed,
+{
+    let tansq = slope.powi(2);
+    let cossq = T::one() / (T::one() + tansq);
+    let sinsq = T::one() - cossq;
+    let mut cos = T::zero();
+    let mut sin;
+    let pnext = poly.exterior.0[poly.next_vertex(idx)];
+    let pprev = poly.exterior.0[poly.prev_vertex(idx)];
+    let clockwise = cross_prod(&pprev, p, &pnext) < T::zero();
+    let slope_prev;
+    let slope_next;
+    // Slope isn't 0, things are complicated
+    if *slope != T::zero() {
+        cos = cossq.sqrt();
+        sin = sinsq.sqrt();
+        if pnext.x() > p.x() {
+            if pprev.x() > p.x() {
+                if pprev.y() >= p.y() && pnext.y() >= p.y() {
+                    if *slope > T::zero() {
+                        slope_prev = (pprev.y() - p.y()) / (pprev.x() - p.x());
+                        if clockwise && *slope <= slope_prev || !clockwise && *slope >= slope_prev {
+                            cos = -cos;
+                            sin = -sin;
+                        } else if clockwise {
+                            cos = -cos;
+                        } else {
+                            sin = -sin;
+                        }
+                    }
+                } else if pprev.y() <= p.y() && pnext.y() <= p.y() {
+                    if *slope > T::zero() {
+                        if !clockwise {
+                            cos = -cos;
+                            sin = -sin;
+                        }
+                    } else {
+                        slope_prev = (pprev.y() - p.y()) / (pprev.x() - p.x());
+                        slope_next = (pnext.y() - p.y()) / (pnext.x() - p.x());
+                        if clockwise {
+                            if *slope <= slope_prev {
+                                cos = -cos;
+                            } else {
+                                sin = -sin;
+                            }
+                        } else if *slope <= slope_next {
+                            sin = -sin;
+                        } else {
+                            cos = -cos;
+                        }
+                    }
+                } else if *slope > T::zero() {
+                    if !clockwise {
+                        cos = -cos;
+                        sin = -sin;
+                    }
+                } else if clockwise {
+                    cos = -cos;
+                } else {
+                    sin = -sin;
+                }
+            } else if *slope < T::zero() {
+                sin = -sin;
+            }
+        } else if pnext.x() < p.x() {
+            if pprev.x() < p.x() {
+                if pprev.y() >= p.y() && pnext.y() >= p.y() {
+                    if *slope > T::zero() {
+                        if clockwise {
+                            cos = -cos;
+                            sin = -sin;
+                        }
+                    } else {
+                        slope_prev = (p.y() - pprev.y()) / (p.x() - pprev.x());
+                        slope_next = (p.y() - pnext.y()) / (p.x() - pnext.x());
+                        if clockwise {
+                            if *slope <= slope_prev {
+                                sin = -sin;
+                            } else {
+                                cos = -cos;
+                            }
+                        } else if *slope <= slope_next {
+                            cos = -cos;
+                        } else {
+                            sin = -sin;
+                        }
+                    }
+                } else if pprev.y() <= p.y() && pnext.y() <= p.y() {
+                    if *slope > T::zero() {
+                        slope_next = (p.y() - pnext.y()) / (p.x() - pnext.x());
+                        if *slope >= slope_next {
+                            cos = -cos;
+                            sin = -sin;
+                        }
+                    } else if clockwise {
+                        sin = -sin;
+                    } else {
+                        cos = -cos;
+                    }
+                } else if *slope > T::zero() {
+                    if clockwise {
+                        cos = -cos;
+                        sin = -sin;
+                    }
+                } else if clockwise {
+                    sin = -sin;
+                } else {
+                    cos = -cos;
+                }
+            } else {
+                //pprev.x() >= p.x()
+                cos = -cos;
+                if *slope > T::zero() {
+                    sin = -sin;
+                }
+            }
+        } else if pprev.x() > p.x() {
+            cos = -cos;
+            if *slope > T::zero() {
+                sin = -sin;
+            }
+        } else if *slope < T::zero() {
+            sin = -sin;
+        }
+    } else {
+        // Slope is 0, things are fairly simple
+        sin = T::zero();
+        if pnext.x() > p.x() {
+            cos = T::one();
+        } else if pnext.x() < p.x() {
+            cos = -T::one();
+        } else if pnext.x() == p.x() {
+            if pprev.x() < p.x() {
+                cos = T::one();
+            } else {
+                cos = -T::one();
+            }
+        }
+    }
+    Point::new(
+        p.x() + T::from(100).unwrap() * cos,
+        p.y() + T::from(100).unwrap() * sin,
+    )
+}
+
+// Perpendicular unit vector of a vertex and a unit vector
+fn unitpvector<T>(p: &Point<T>, u: &Point<T>) -> Point<T>
+where
+    T: Float,
+{
+    let hundred = T::from(100).unwrap();
+    let vertical;
+    let mut slope;
+    slope = T::zero();
+    if p.x() == u.x() {
+        vertical = true;
+    } else {
+        vertical = false;
+    }
+    if !vertical {
+        if p.y() == u.y() {
+            slope = T::zero();
+        } else if u.x() > p.x() {
+            slope = (u.y() - p.y()) / (u.x() - p.x());
+        } else {
+            slope = (p.y() - u.y()) / (p.x() - u.x());
+        }
+    }
+    let upx;
+    let upy;
+    if vertical {
+        upy = p.y();
+        if u.y() > p.y() {
+            upx = p.x() + hundred;
+        } else {
+            upx = p.x() - hundred;
+        }
+        Point::new(upx, upy)
+    } else if slope == T::zero() {
+        upx = p.x();
+        if u.x() > p.x() {
+            upy = p.y() - hundred;
+        } else {
+            upy = p.y() + hundred;
+        }
+        Point::new(upx, upy)
+    } else {
+        // Not a special case
+        let sperp = -T::one() / slope;
+        let tansq = sperp * sperp;
+        let cossq = T::one() / (T::one() + tansq);
+        let sinsq = T::one() - cossq;
+        let mut cos = cossq.sqrt();
+        let mut sin = sinsq.sqrt();
+        if u.x() > p.x() {
+            sin = -sin;
+            if slope < T::zero() {
+                cos = -cos;
+            }
+        } else if slope > T::zero() {
+            cos = -cos;
+        }
+        Point::new(p.x() + hundred * cos, p.y() + hundred * sin)
+    }
+}
+
+// Angle between a vertex and an edge
+fn vertex_line_angle<T>(poly: &Polygon<T>, p: &Point<T>, m: &T, vertical: bool, idx: &usize) -> T
+where
+    T: Float + FloatConst + Signed,
+{
+    let hundred = T::from(100).unwrap();
+    let pnext = poly.exterior.0[poly.next_vertex(idx)];
+    let pprev = poly.exterior.0[poly.prev_vertex(idx)];
+    let clockwise = cross_prod(&pprev, p, &pnext) < T::zero();
+    let punit;
+    if !vertical {
+        punit = unitvector(m, poly, p, idx);
+    } else {
+        match clockwise {
+            true => {
+                if p.x() > pprev.x() {
+                    punit = Point::new(p.x(), p.y() - hundred);
+                } else if p.x() == pprev.x() {
+                    if p.y() > pprev.y() {
+                        punit = Point::new(p.x(), p.y() + hundred);
+                    } else {
+                        // implies p.y() < pprev.y()
+                        // it's safe not to explicitly cover p.y() == pprev.y() because that
+                        // implies that the x values are equal, and the y values are equal,
+                        // and this is impossible
+                        punit = Point::new(p.x(), p.y() - hundred);
+                    }
+                } else {
+                    // implies p.x() < pprev.x()
+                    punit = Point::new(p.x(), p.y() + hundred);
+                }
+            }
+            false => {
+                if p.x() > pprev.x() {
+                    punit = Point::new(p.x(), p.y() + hundred);
+                } else if p.x() == pprev.x() {
+                    if p.y() > pprev.y() {
+                        punit = Point::new(p.x(), p.y() + hundred);
+                    } else {
+                        // implies p.y() < pprev.y()
+                        // it's safe not to explicitly cover p.y() == pprev.y() because that
+                        // implies that the x values are equal, and the y values are equal,
+                        // and this is impossible
+                        punit = Point::new(p.x(), p.y() - hundred);
+                    }
+                } else {
+                    // implies p.x() < pprev.x()
+                    punit = Point::new(p.x(), p.y() - hundred);
+                }
+            }
+        }
+    }
+    let triarea = triangle_area(p, &punit, &pnext);
+    let edgelen = p.distance(&pnext);
+    let mut sine = triarea / (T::from(0.5).unwrap() * T::from(100).unwrap() * edgelen);
+    if sine < -T::one() {
+        sine = T::one();
+    }
+    if sine > T::one() {
+        sine = T::one();
+    }
+    let angle;
+    let perpunit = unitpvector(p, &punit);
+    let mut obtuse = false;
+    let left = leftturn(p, &perpunit, &pnext);
+    if clockwise {
+        if left == 0 {
+            obtuse = true;
+        }
+        if left == -1 {
+            angle = T::PI() / (T::one() + T::one());
+        } else if !obtuse {
+            angle = (-sine).asin();
+        } else {
+            angle = T::PI() - (-sine).asin();
+        }
+    } else {
+        if left == 0 {
+            obtuse = true;
+        }
+        if left == -1 {
+            angle = T::PI() / (T::one() + T::one());
+        } else if !obtuse {
+            angle = sine.asin();
+        } else {
+            angle = T::PI() - sine.asin();
+        }
+    }
+    angle
+}
+
+// self-explanatory
+fn triangle_area<T>(a: &Point<T>, b: &Point<T>, c: &Point<T>) -> T
+where
+    T: Float,
+{
+    (T::from(0.5).unwrap() * (a.x() * b.y() - a.y() * b.x() + a.y() * c.x() - a.x() * c.y() + b.x() * c.y() - c.x() * b.y()))
+}
+
+// positive implies a -> b -> c is counter-clockwise, negative implies clockwise
+fn cross_prod<T>(p_a: &Point<T>, p_b: &Point<T>, p_c: &Point<T>) -> T
+where
+    T: Float,
+{
+    (p_b.x() - p_a.x()) * (p_c.y() - p_a.y()) - (p_b.y() - p_a.y()) * (p_c.x() - p_a.x())
+}
+
+// Does abc turn left?
+fn leftturn<T>(a: &Point<T>, b: &Point<T>, c: &Point<T>) -> i8
+where
+    T: Float,
+{
+    let narea = triangle_area(a, b, c);
+    if narea > T::zero() {
+        1
+    } else if narea < T::zero() {
+        0
+    } else {
+        -1
+    }
+}
+
+// Calculate next set of caliper points
+fn nextpoints<T>(state: &mut Polydist<T>)
+where
+    T: Float + FloatConst + Signed,
+{
+    state.alignment = None;
+    state.ip1 = false;
+    state.iq2 = false;
+    state.ap1 = vertex_line_angle(
+        state.poly1,
+        &state.p1,
+        &state.slope,
+        state.vertical,
+        &state.p1_idx,
+    );
+    state.aq2 = vertex_line_angle(
+        state.poly2,
+        &state.q2,
+        &state.slope,
+        state.vertical,
+        &state.q2_idx,
+    );
+    let minangle = state.ap1.min(state.aq2);
+    state.p1prev = state.p1;
+    state.p1next = state.p1prev;
+    state.q2prev = state.q2;
+    state.q2next = state.q2prev;
+    // iff (ap1 - minangle) is less than epsilon, alignment is edge-vertex (P-Q)
+    // iff (aq2 - minangle) is less than epsilon, alignment is edge-vertex (Q-P)
+    // if both are within epsilon, alignment is edge-edge
+    // in each of the above, we also have to check for overlap, and in the case of
+    // edge-edge alignment, additional cases must be considered.
+    //
+    // assume the calipers are rotated θ degrees around pi and qj, and that
+    // we have hit vertex q` and edge [p`, p^]
+    // check whether there exists a line segment [p, p*] which is orthogonal to [qj, q`]
+    // compute the intersection of lines [pi, p*] and [qj, q`]
+    // if this intersection q† exists, and ≠ qj or q`, compute the distance
+    // between pi and q†, and compare it to the current minimum.
+    // If the calipers intersect with edges on both polygons (implying the edges are parallel),
+    // intersections must be computed between both segments, and if one is
+    // found, the [pi, p`] - [qj, q`] edge-edge orthogonal distance is found and compared.
+    // see Pirzadeh (1999), p31
+    if (state.ap1 - minangle).abs() < T::from(0.002).unwrap() {
+        state.ip1 = true;
+        let p1next = state.poly1.next_vertex(&state.p1_idx);
+        state.p1next = state.poly1.exterior.0[p1next];
+        state.p1_idx = p1next;
+        state.alignment = Some(Aligned::EdgeVertexP);
+    }
+    if (state.aq2 - minangle).abs() < T::from(0.002).unwrap() {
+        state.iq2 = true;
+        let q2next = state.poly2.next_vertex(&state.q2_idx);
+        state.q2next = state.poly2.exterior.0[q2next];
+        state.q2_idx = q2next;
+        state.alignment = match state.alignment {
+            None => Some(Aligned::EdgeVertexQ),
+            Some(_) => Some(Aligned::EdgeEdge),
+        }
+    }
+    if state.ip1 {
+        match state.p1.x() == state.p1next.x() {
+            // The P line of support is vertical
+            true => {
+                state.vertical = true;
+                state.slope = T::zero();
+            }
+            false => {
+                state.vertical = false;
+                if state.p1.x() > state.p1next.x() {
+                    state.slope = (state.p1.y() - state.p1next.y()) / (state.p1.x() - state.p1next.x());
+                } else {
+                    state.slope = (state.p1next.y() - state.p1.y()) / (state.p1next.x() - state.p1.x());
+                }
+            }
+        }
+    }
+    if state.iq2 {
+        match state.q2.x() == state.q2next.x() {
+            true => {
+                // The Q line of support is vertical
+                state.vertical = true;
+                state.slope = T::zero();
+            }
+            false => {
+                state.vertical = false;
+                if state.q2.x() > state.q2next.x() {
+                    state.slope = (state.q2.y() - state.q2next.y()) / (state.q2.x() - state.q2next.x());
+                } else {
+                    state.slope = (state.q2next.y() - state.q2.y()) / (state.q2next.x() - state.q2.x());
+                }
+            }
+        }
+    }
+    // A start value's been set, and both polygon indices are in their initial
+    // positions -- we're finished, so return the minimum distance
+    if state.p1 == state.ymin1 && state.q2 == state.ymax2 && !state.start.is_none() {
+        state.finished = true;
+    } else {
+        state.start = Some(false);
+        state.p1 = state.p1next;
+        state.q2 = state.q2next;
+    }
+}
+
+// compute the minimum distance between entities (edges or vertices)
+fn computemin<T>(state: &mut Polydist<T>)
+where
+    T: Float + Signed,
+{
+    let u;
+    let u1;
+    let u2;
+    let mut newdist = state.p1.distance(&state.q2);
+    if newdist <= state.dist {
+        // New minimum distance is between p1 and q2
+        state.dist = newdist;
+    }
+    match state.alignment {
+        Some(Aligned::EdgeVertexP) => {
+            // one line of support coincides with a vertex on Q, the other with an edge on P
+            if !state.vertical {
+                if state.slope != T::zero() {
+                    u = unitvector(
+                        &(-T::one() / state.slope),
+                        state.poly2,
+                        &state.q2,
+                        &state.q2_idx,
+                    );
+                } else {
+                    u = Point::new(state.q2.x(), state.q2.y() + T::from(100).unwrap());
+                }
+            } else {
+                u = unitvector(&T::zero(), state.poly2, &state.q2, &state.q2_idx);
+            }
+            let line_1 = leftturn(&u, &state.q2, &state.p1);
+            let line_2 = leftturn(&u, &state.q2, &state.p1prev);
+            if line_1 != line_2 && line_1 != -1 && line_2 != -1 {
+                // an orthogonal intersection exists
+                newdist = state.q2.vertex_line_distance(&state.p1prev, &state.p1);
+                if newdist <= state.dist {
+                    // New minimum distance is between edge (p1prev, p1) and q2
+                    state.dist = newdist;
+                }
+            }
+        }
+        Some(Aligned::EdgeVertexQ) => {
+            // one line of support coincides with a vertex on P, the other with an edge on Q
+            if !state.vertical {
+                if state.slope != T::zero() {
+                    u = unitvector(
+                        &(-T::one() / state.slope),
+                        state.poly1,
+                        &state.p1,
+                        &state.p1_idx,
+                    );
+                } else {
+                    u = Point::new(state.p1.x(), state.p1.y() + T::from(100).unwrap());
+                }
+            } else {
+                u = unitvector(&T::zero(), state.poly1, &state.p1, &state.p1_idx);
+            }
+            let line_1 = leftturn(&u, &state.p1, &state.q2);
+            let line_2 = leftturn(&u, &state.p1, &state.q2prev);
+            if line_1 != line_2 && line_1 != -1 && line_2 != -1 {
+                // an orthogonal intersection exists
+                newdist = state.p1.vertex_line_distance(&state.q2prev, &state.q2);
+                if newdist <= state.dist {
+                    // New minimum distance is between edge (q2prev, q2) and p1
+                    state.dist = newdist;
+                }
+            }
+        }
+        Some(Aligned::EdgeEdge) => {
+            // both lines of support coincide with edges (i.e. they're parallel)
+            newdist = state.p1.distance(&state.q2prev);
+            if newdist <= state.dist {
+                // New minimum distance is between p1 and q2prev
+                state.dist = newdist;
+            }
+            newdist = state.p1prev.distance(&state.q2);
+            if newdist <= state.dist {
+                // New minimum distance is between p1prev and q2
+                state.dist = newdist;
+            }
+            if !state.vertical {
+                if state.slope != T::zero() {
+                    u1 = unitvector(
+                        &(-T::one() / state.slope),
+                        state.poly1,
+                        &state.p1prev,
+                        &state.p1_idx,
+                    );
+                    u2 = unitvector(
+                        &(-T::one() / state.slope),
+                        state.poly1,
+                        &state.p1,
+                        &state.p1_idx,
+                    );
+                } else {
+                    u1 = Point::new(state.p1prev.x(), state.p1prev.y() + T::from(100).unwrap());
+                    u2 = Point::new(state.p1.x(), state.p1.y() + T::from(100).unwrap());
+                }
+            } else {
+                u1 = unitvector(&T::zero(), state.poly1, &state.p1prev, &state.p1_idx);
+                u2 = unitvector(&T::zero(), state.poly1, &state.p1, &state.p1_idx);
+            }
+            let line_1a = leftturn(&u1, &state.p1prev, &state.q2prev);
+            let line_1b = leftturn(&u1, &state.p1prev, &state.q2);
+            let line_2a = leftturn(&u2, &state.p1, &state.q2prev);
+            let line_2b = leftturn(&u2, &state.p1, &state.q2);
+            if line_1a != line_1b && line_1a != -1 && line_1b != -1 || line_2a != line_2b && line_2a != -1 && line_2b != -2 {
+                // an orthogonal intersection exists
+                newdist = state.p1.vertex_line_distance(&state.q2prev, &state.q2);
+                if newdist <= state.dist {
+                    // New minimum distance is between edge (p1prev, p1) and q2prev
+                    state.dist = newdist;
+                }
+            }
+        }
+        _ => unreachable!(),
+    }
+}
